+++ src/components/Site/SiteNav.react.js
@@ -51,18 +51,20 @@ const SiteNav = ({
   rightColumnComponent,
   collapse = true,
   routerContextComponentType,
+  stickyTop = false
 }: Props): React.Node => {
-  const classes = cn("header d-lg-flex p-0", { collapse });
+  const sticky = stickyTop ? "sticky-top" : ""
+  const classes = cn("header d-lg-flex p-0 " + sticky, { collapse });
   return (
     <div className={classes}>
       <Container>
         {children || (
           <Grid.Row className="align-items-center">
-            <Grid.Col lg={3} className="ml-auto" ignoreCol={true}>
+            {/* <Grid.Col lg={3} className="ml-auto" ignoreCol={true}> */}
               {/* @TODO: add InlineSearchForm  */}
               {/* {rightColumnComponent || (withSearchForm && <InlineSearchForm />)} */}
-              {rightColumnComponent}
-            </Grid.Col>
+              {/* {rightColumnComponent} */}
+            {/* </Grid.Col> */}
             <Grid.Col className="col-lg order-lg-first">
               <Nav
                 tabbed